requirements.txt
docling
llama-index
llama-index-embeddings-openai
llama-index-vector-stores-pinecone
pinecone
openai
python-dotenv
pymongo